更换了创作方法使用数据库开始,以页面

+++ database/migrations/2025_03_18_114503_create_login_logs_table.php
@@ -12,9 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('login_logs', function (Blueprint $table) {
-            $table->id();
-            // 修改为 unsignedInteger 类型
-            $table->unsignedInteger('user_id')->comment('用户 ID');
+            $table->id(); // 这是自增主键
+            $table->unsignedBigInteger('user_id')->comment('用户 ID');
             $table->string('ip_address', 45)->comment('登录 IP 地址');
             $table->timestamp('login_time')->comment('登录时间');
             $table->timestamps();
